Fix: Restore server-side AI orchestration architecture
- I restored missing server-side files critical for AI orchestration:
  - `server/schemas.ts` (AI response schema definitions)
  - `server/prompts.ts` (prompt generation functions and `genrePersonaDetails`)
  - `server/utils.ts` (parsing, slugify, AI constants, image helpers, error handlers)

- I refactored `services/geminiService.ts` to be a thin client:
  - I removed all client-side prompt construction, AI SDK calls, and response processing.
  - I modified service functions to call the appropriate `/api/adventure/...` server orchestration endpoints.
  - The client now expects final, structured data from the server.

- I verified that `server/routes/adventureApi.ts` correctly implements the orchestration logic using the restored server files.
- I confirmed type correctness for both client and server codebases.
- I ensured the project builds successfully.

This resolves the issue where the application did not reflect the intended server-side handling of AI interactions. Unit test execution remains blocked by environment issues.
